WIP switch to standard HMRC HTTP error representation
