--- app/uk/gov/hmrc/taxhistory/connectors/BaseConnector.scala
@@ -1,58 +1,0 @@
-/*
- * Copyright 2018 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.taxhistory.connectors
-
-import play.api.http.Status
-import uk.gov.hmrc.domain.Nino
-import uk.gov.hmrc.http._
-import uk.gov.hmrc.play.config.{AppName, ServicesConfig}
-import uk.gov.hmrc.play.config.ServicesConfig
-import uk.gov.hmrc.play.http.logging.MdcLoggingExecutionContext._
-import uk.gov.hmrc.taxhistory.auditable.Auditable
-import uk.gov.hmrc.taxhistory.metrics.{MetricsEnum, TaxHistoryMetrics}
-import uk.gov.hmrc.taxhistory.utils.TaxHistoryLogger
-
-import scala.concurrent.Future
-
-trait BaseConnector extends AnyRef with TaxHistoryLogger {
-
-  val servicesConfig: ServicesConfig
-
-  def httpGet: CoreGet
-
-  def httpPost: CorePost
-
-  def originatorId: String
-
-  def metrics: TaxHistoryMetrics = TaxHistoryMetrics
-
-  val defaultVersion: Int = -1
-
-  implicit val httpReads: HttpReads[HttpResponse] = new HttpReads[HttpResponse] {
-    override def read(method: String, url: String, response: HttpResponse) = response
-  }
-
-  def basicNpsHeaders(hc: HeaderCarrier): HeaderCarrier = {
-    hc.withExtraHeaders("Gov-Uk-Originator-Id" -> originatorId)
-  }
-
-  def withoutSuffix(nino: Nino): String = {
-    val BASIC_NINO_LENGTH = 8
-    nino.value.take(BASIC_NINO_LENGTH)
-  }
-
-}
